+++ CHANGELOG.md
@@ -11,10 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Peer to Peer bad_actor sharing, e.g. it gets saved on the DHT and consumed by other peers
 - RESTful API numbers resource (`/numbers`) now only returns numbers like +441234567890 or 441234567890
 - RESTful API number resource (`/numbers/{number}`) now only accepts +441234567890 or 441234567890 formats
+- Generate a SentryPeer node ID at startup to be used in all `bad_actor` events
+- Stop the SIP daemon thread correctly on shutdown
 
 ### Tests
 - More configuration CLI options tested
 - Bad Actor to JSON conversion tests - used to save on the DHT
+- Memory cleanups on tests
 
 ## [1.2.0] - 2022-02-25
 - Add signal handlers to gracefully shut down the SIP, P2P and HTTP daemons